+++ packages/snippets/src/acro-dynamic-form/src/index.vue
@@ -1,5 +1,5 @@
 <template>
-  <Form ref="formRef" :model="model" v-bind="attrs">
+  <Form ref="formRef" :model="model" v-bind="formProps">
     <!-- 动态组件渲染 -->
     <RenderFormItem v-for="(field, index) in fields" :key="index" :field="field">
       <RenderComponent :field="field" :model="model" :slots="slots" />
@@ -20,8 +20,10 @@
 <script lang="tsx" setup>
   import type { FormInstance } from '@arco-design/web-vue';
   import { Button, Form, FormItem, Space } from '@arco-design/web-vue';
-  import { reactive, ref, toRef, useAttrs } from 'vue';
+  import { computed, reactive, ref, toRef, useAttrs } from 'vue';
   import { toReactive } from '@vueuse/core';
+  import { deepMerge } from 'mixte';
+  import { pick } from 'radash';
   import RenderFormItem from './components/RenderFormItem.vue';
   import RenderComponent from './components/RenderComponent.vue';
   import { useActionButtonArea } from './composables/useActionButtonArea';
@@ -46,6 +48,16 @@
   props.fields?.forEach((field) => {
     if (field.defaultValue != null)
       model[field.field] = model[field.field] ?? field.defaultValue;
+  });
+
+  const formProps = computed(() => {
+    return deepMerge(
+      pick(props, [
+        'layout', 'size', 'labelColProps', 'wrapperColProps', 'labelAlign', // eslint-disable-line antfu/consistent-list-newline
+        'disabled', 'rules', 'autoLabelWidth', 'id', 'scrollToFirstError', // eslint-disable-line antfu/consistent-list-newline
+      ]),
+      attrs,
+    );
   });
 
   const {